Move convert_image to utils
