+++ wgp/__init__.py
@@ -1,0 +1,5 @@
+"""Utilities for the Wan2GP project."""
+
+from . import utils
+
+__all__ = ["utils"]
